Importing new codes
With the creation of 0_Createdirs.py, it is possible to create the directories to where the data will be processed.
The creation of 3rd_ConversionGeojsonToShapefile.py was done to convert from Geojson format file into shapefile. This py file uses de OGR2OGR command to process geospatial data.

Those codes are part of MaxarBuildingFootprints extraction and geospatial data handling.

The '3rd' code is in development. However it can be used successfully.